Added MultiDoc support
Most backends do not parse multiple files and therefore do not need
to be passed all unsaved files in an editor. Now only backends that support
the org.gnome.CodeAssist.MultiDoc service will receive unsaved documents
outside of the primary document. This is currently only implemented in the
C backend.