python: fix pylint
in some recent pylint version, the output was changed to include
sapartor lines made out of minus signs.

the pylint code in gnome-code-assistance assumed any line returned
from pylint would have ":" in it, and that change in pylint broke
that assumption.

with this commit, we only process pylint lines if they have ":"
in them, which means we ignore any output that doesn't contain
diagnostics.